requirements.txt
openai>=1.0.0
supabase>=2.0.0
python-dotenv
beautifulsoup4
snscrape
feedparser
googlemaps
